app-python/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo==4.7.3
google-cloud-pubsub
